making thread changes for running GUI on mac

+++ include/Feature.hpp
@@ -10,11 +10,11 @@ class Feature {
     public:
         typedef std::shared_ptr<Feature> ptr;
         unsigned long featureID;
-        cv::Point2i kp;
+        cv::Point2d kp;
         std::weak_ptr<Frame> frame;
         std::weak_ptr<MapPoint> mapPoint;
         Feature(){}
-        Feature(std::shared_ptr<Frame> _frame, const cv::Point2i &_kp) : frame(_frame), kp(_kp) {}
+        Feature(std::shared_ptr<Frame> _frame, const cv::Point2d &_kp) : frame(_frame), kp(_kp) {}
 };
